Interface to conda

R functions for managing Python conda environments.

conda_list(conda = "auto")

conda_create(
  envname = NULL,
  packages = "python",
  forge = TRUE,
  channel = character(),
  conda = "auto"
)

conda_remove(envname, packages = NULL, conda = "auto")

conda_install(
  envname = NULL,
  packages,
  forge = TRUE,
  channel = character(),
  pip = FALSE,
  pip_options = character(),
  pip_ignore_installed = FALSE,
  conda = "auto",
  python_version = NULL,
  ...
)

conda_binary(conda = "auto")

conda_version(conda = "auto")

conda_python(envname = NULL, conda = "auto")

Arguments

conda
The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda for more details.

envname
The name of, or path to, a conda environment.

packages
A character vector, indicating package names which should be installed or removed.

forge
Boolean; include the Conda Forge repository?

channel
An optional character vector of Conda channels to include. When specified, the forge argument is ignored. If you need to specify multiple channels, including the Conda Forge, you can use c("conda-forge", <other channels>).

pip
Boolean; use pip when downloading or installing packages? Defaults to FALSE.

pip_options
An optional character vector of additional command line arguments to be passed to pip. Only relevant when pip = TRUE.

pip_ignore_installed
Ignore installed versions when using pip. This is TRUE by default so that specific package versions can be installed even if they are downgrades. The FALSE option is useful for situations where you don't want a pip install to attempt an overwrite of a conda binary package (e.g. SciPy on Windows which is very difficult to install via pip due to compilation requirements).

...
Optional arguments, reserved for future expansion.

Value

conda_list() returns a data frame with the names and paths to the respective python binaries of available environments. conda_create() returns the path to the python binary of the created environment. conda_binary() returns the location of the main conda binary or NULL if none can be found.

Finding Conda

When conda = "auto", reticulate will attempt to automatically find an Anaconda / Miniconda installation and use that. reticulate will search the following locations:

The location specified by the reticulate.conda_binary R option;

The miniconda_path() location (if it exists);

The program PATH;

A set of pre-defined locations where Conda is typically installed.