List Conda Environments

List all of the available conda environments on the system.

conda_list(conda = "auto")

Arguments

conda

The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See conda_binary() for more details on how reticulate tries to resolve the conda executable.

Value

An Rdata.frame, with name giving the name of the associated environment, and python giving the path to the Python binary associated with that environment.

Details

Environments available are listed as by:

conda info --json

Note that this function will report all available conda environments on the system, even those associated with a different conda installation than the requested conda binary.

See also