miniconda-params {reticulate}	R Documentation

miniconda-params

Description

miniconda-params

Arguments

path
The location where Miniconda is (or should be) installed. Note that the Miniconda installer does not support paths containing spaces. See miniconda_path for more details on the default path used by reticulate.

[Package reticulate version 1.25 Index]