miniconda_update {reticulate}	R Documentation

Update Miniconda

Description

Update Miniconda to the latest version.

Usage

miniconda_update(path = miniconda_path())

Arguments

path
The location where Miniconda is (or should be) installed. Note that the Miniconda installer does not support paths containing spaces. See miniconda_path for more details on the default path used by reticulate.

See Also

Other miniconda-tools: install_miniconda(), miniconda_uninstall()

[Package reticulate version 1.25 Index]