py {reticulate}	R Documentation

Interact with the Python Main Module

Description

The py object provides a means for interacting with the Python main session directly from R. Python objects accessed through py are automatically converted into R objects, and can be used with any other R functions as needed.

Usage

py

Format

An R object acting as an interface to the Python main module.

[Package reticulate version 1.25 Index]