py_config {reticulate}	R Documentation

Python configuration

Description

Retrieve information about the version of Python currently being used by reticulate.

Usage

py_config()

Details

If Python has not yet been initialized, then calling py_config() will force the initialization of Python. See py_discover_config() for more details.

Value

Information about the version of Python in use, as an R list with class "py_config".

[Package reticulate version 1.25 Index]