py_discover_config {reticulate}	R Documentation

Discover the version of Python to use with reticulate.

Description

This function enables callers to check which versions of Python will be discovered on a system as well as which one will be chosen for use with reticulate.

Usage

py_discover_config(required_module = NULL, use_environment = NULL)

Arguments

required_module
A optional module name that must be available in order for a version of Python to be used.

use_environment
An optional virtual/conda environment name to prefer in the search.

Value

Python configuration object.

[Package reticulate version 1.25 Index]