py_has_attr {reticulate}	R Documentation

Check if a Python object has an attribute

Description

Check whether a Python object x has an attribute name.

Usage

py_has_attr(x, name)

Arguments

x
A python object.

name
The attribute to be accessed.

Value

TRUE if the object has the attribute name, and FALSE otherwise.

[Package reticulate version 1.25 Index]