py_id {reticulate}	R Documentation

Unique identifer for Python object

Description

Get a globally unique identifer for a Python object.

Usage

py_id(object)

Arguments

object
Python object

Value

Unique identifer (as integer) or NULL

Note

In the current implementation of CPython this is the memory address of the object.

[Package reticulate version 1.25 Index]