py_main_thread_func {reticulate}	R Documentation

Create a Python function that will always be called on the main thread

Description

This function is helpful when you need to provide a callback to a Python library which may invoke the callback on a background thread. As R functions must run on the main thread, wrapping the R function with py_main_thread_func() will ensure that R code is only executed on the main thread.

Usage

py_main_thread_func(f)

Arguments

f
An R function with arbitrary arguments

Value

A Python function that delegates to the passed R function, which is guaranteed to always be called on the main thread.

[Package reticulate version 1.25 Index]