py_to_r_wrapper {reticulate}	R Documentation

R wrapper for Python objects

Description

S3 method to create a custom R wrapper for a Python object. The default wrapper is either an R environment or an R function (for callable python objects).

Usage

py_to_r_wrapper(x)

Arguments

x
Python object

[Package reticulate version 1.25 Index]