py_version {reticulate}	R Documentation

Python version

Description

Get the version of Python currently being used by reticulate.

Usage

py_version()

Value

The version of Python currently used, or NULL if Python has not yet been initialized by reticulate.

[Package reticulate version 1.25 Index]