py_versions_windows {reticulate}	R Documentation

Discover versions of Python installed on a Windows system

Description

Discover versions of Python installed on a Windows system

Usage

py_versions_windows()

Value

Data frame with type, hive, install_path, executable_path, and version.

[Package reticulate version 1.25 Index]