R Interface to Python

R interface to Python modules, classes, and functions. When calling into Python R data types are automatically converted to their equivalent Python types. When values are returned from Python to R they are converted back to R types. The reticulate package is compatible with all versions of Python >= 3.6. Integration with NumPy requires NumPy version 1.6 or higher.

See also

Author

Maintainer: Tomasz Kalinowski tomasz@posit.co [contributor]

Authors:

Kevin Ushey kevin@posit.co

JJ Allaire jj@posit.co

Yuan Tang terrytangyuan@gmail.com (ORCID) [copyright holder]

Other contributors:

RStudio [copyright holder, funder]

Dirk Eddelbuettel edd@debian.org [contributor, copyright holder]

Bryan Lewis blewis@illposed.net [contributor, copyright holder]

Sigrid Keydana sigrid@posit.co [contributor]

Ryan Hafen rhafen@gmail.com [contributor, copyright holder]

Marcus Geelnard (TinyThread library, http://tinythreadpp.bitsnbites.eu/) [contributor, copyright holder]