install_miniconda {reticulate}	R Documentation

Install Miniconda

Description

Download the Miniconda installer, and use it to install Miniconda.

Usage

install_miniconda(path = miniconda_path(), update = TRUE, force = FALSE)

Arguments

path
The location where Miniconda is (or should be) installed. Note that the Miniconda installer does not support paths containing spaces. See miniconda_path for more details on the default path used by reticulate.

update
Boolean; update to the latest version of Miniconda after installation?

force
Boolean; force re-installation if Miniconda is already installed at the requested path?

Details

For arm64 builds of R on macOS, install_miniconda() will use binaries from miniforge instead.

Note

If you encounter binary incompatibilities between R and Miniconda, a scripted build and installation of Python from sources can be performed by install_python()

See Also

Other miniconda-tools: miniconda_uninstall(), miniconda_update()

[Package reticulate version 1.25 Index]