install_python {reticulate}	R Documentation

Install Python

Description

Download and install Python, using the pyenv. and pyenv-win projects.

Usage

install_python(version = "3.9:latest", list = FALSE, force = FALSE)

Arguments

version
The version of Python to install.

list
Boolean; if set, list the set of available Python versions?

force
Boolean; force re-installation even if the requested version of Python is already installed?

Details

In general, it is recommended that Python virtual environments are created using the copies of Python installed by install_python(). For example:

library(reticulate)
version <- "3.9.12"
install_python(version)
virtualenv_create("my-environment", version = version)
use_virtualenv("my-environment")

# There is also support for a ":latest" suffix to select the latest patch release
install_python("3.9:latest") # install latest patch available at python.org

# select the latest 3.9.* patch installed locally
virtualenv_create("my-environment", version = "3.9:latest")

Note

On macOS and Linux this will build Python from sources, which may take a few minutes.

[Package reticulate version 1.25 Index]