with.python.builtin.object {reticulate}	R Documentation

Evaluate an expression within a context.

Description

The with method for objects of type python.builtin.object implements the context manager protocol used by the Python with statement. The passed object must implement the context manager (__enter__ and __exit__ methods.

Usage

## S3 method for class 'python.builtin.object'
with(data, expr, as = NULL, ...)

Arguments

data
Context to enter and exit

expr
Expression to evaluate within the context

as
Name of variable to assign context to for the duration of the expression's evaluation (optional).

...
Unused

[Package reticulate version 1.25 Index]